refactor: make AddressExtensions helpers null-safe and validate memo inputs
Return nullable SendSrc? from firstSendSrc/findCurrentSrc so missing
addresses/accounts/tokens don't throw NoSuchElementException. Collapses
the duplicate definitions that lived at the bottom of SwapFormViewModel
into the shared helpers in the send package. Adds require() preconditions
to rujiRewardsMemo and tcyUnstakeMemo so malformed inputs can't produce
invalid THORChain memos.
