+++ app/src/main/java/com/vultisig/wallet/ui/models/swap/SwapFormViewModel.kt
@@ -35,6 +35,8 @@ import com.vultisig.wallet.ui.models.mappers.FiatValueToStringMapper
 import com.vultisig.wallet.ui.models.send.InvalidTransactionDataException
 import com.vultisig.wallet.ui.models.send.SendSrc
 import com.vultisig.wallet.ui.models.send.TokenBalanceUiModel
+import com.vultisig.wallet.ui.models.send.findCurrentSrc
+import com.vultisig.wallet.ui.models.send.firstSendSrc
 import com.vultisig.wallet.ui.models.swap.SwapTokenSelector.Companion.ARG_SELECTED_DST_TOKEN_ID
 import com.vultisig.wallet.ui.models.swap.SwapTokenSelector.Companion.ARG_SELECTED_SRC_TOKEN_ID
 import com.vultisig.wallet.ui.navigation.Destination
@@ -1018,51 +1020,4 @@ internal fun MutableStateFlow<SendSrc?>.updateSrc(
                 addresses.findCurrentSrc(selectedTokenId, selectedSrcValue)
             }
         }
-}
-
-internal fun List<Address>.firstSendSrc(selectedTokenId: String?, filterByChain: Chain?): SendSrc? {
-    val address =
-        when {
-            !selectedTokenId.isNullOrBlank() ->
-                firstOrNull { it -> it.accounts.any { it.token.id == selectedTokenId } }
-                    ?: run {
-                        Timber.w("selectedTokenId %s not found", selectedTokenId)
-                        this.firstOrNull() // fallback: best-effort, token not found in any address
-                    }
-                    ?: return null
-
-            filterByChain != null -> firstOrNull { it.chain == filterByChain } ?: return null
-            else -> firstOrNull() ?: return null
-        }
-    val account =
-        when {
-            !selectedTokenId.isNullOrBlank() ->
-                address.accounts.firstOrNull { it.token.id == selectedTokenId }
-                    ?: address.accounts.firstOrNull()
-                    ?: return null
-
-            filterByChain != null ->
-                address.accounts.firstOrNull { it.token.isNativeToken } ?: return null
-
-            else -> address.accounts.firstOrNull() ?: return null
-        }
-
-    return SendSrc(address, account)
-}
-
-internal fun List<Address>.findCurrentSrc(selectedTokenId: String?, currentSrc: SendSrc): SendSrc? {
-    if (selectedTokenId == null) {
-        val selectedAddress = currentSrc.address
-        val selectedAccount = currentSrc.account
-        val address =
-            firstOrNull {
-                it.chain == selectedAddress.chain && it.address == selectedAddress.address
-            } ?: return null
-        val account =
-            address.accounts.firstOrNull { it.token.ticker == selectedAccount.token.ticker }
-                ?: return null
-        return SendSrc(address, account)
-    } else {
-        return firstSendSrc(selectedTokenId, null)
-    }
 }
